Create Gradient of Loss Function
Consider a simple neural network with 2 input nodes - x1 and x2 connected to an output node via weights w1 and w2. Thus,

y = w1*x1 + w2*x2

compute the loss function, and the gradient of the loss function with respect to each of the weights w1 and w2. 

Use the table below to test your result

x1   x2    y
1    0      1
2    1      9
0     1     1
-2    1    7